new uno

+++ main.py
@@ -1,9 +1,11 @@
+import glob
 from fastapi import FastAPI, File, UploadFile
 from fastapi.staticfiles import StaticFiles
 from fastapi.responses import JSONResponse
 from fastapi.encoders import jsonable_encoder
 import matplotlib.pyplot as plt
 import os
+from PIL import Image
 import uvicorn
 import base64
 import heartpy as hb
@@ -79,9 +81,12 @@ async def original_signal():
 async def spectrum():
     file_path = "./data/heart.csv"
     ecg, wd, m = setup(file_path)
-    file_image = "./images/spectrum.png"
     
-    os.remove(file_image)
+    file_image = "./images/spectrum.png"
+    image_list = glob.glob(file_image)
+    if image_list:
+        os.remove(file_image)
+    
     plt.title("Frekuensi Spektrum Sinyal Jantung")
     plt.xlim(0,0.6)
     plt.ylim(0, 500)
